pyproject.toml
[project]
name = "a2a-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langgraph-a2a-client>=0.1.5",
    "langgraph-a2a-server>=0.1.6",
    "strands-agents>=1.22.0",
]
